random password generation with given phrase
ever wanted to generate a password that will be hard to crack? 
well here is a code that will generate a random encryption for the alphabet(all lowercase no spaces from the ASCII table, can be altered to be used however you want), you enter a lowercase, no spaces phrase(works up to 9 letters atm when running both encryption and decryption functions, for example, "hello"), then there is two option: encrypt it and decrypt it, the encryption takes the phrase and encrypts it with the random encryption that was generated, the decryption does the opposite and gives back the original phrase.