Update EncryptionAndDecryptionJava
ever wanted to generate a password that will be hard to crack? 
well here is a code that will generate a random encryption for the alphabet(all lowercase no spaces from the ASCII table, where each letter has a value that is between 2-5 characters which can be changed, the code may be altered to be used however you want), you enter a lowercase, no spaces, phrase(for example: "mynameisjeff"), then there are two options: encrypt the phrase or/and decrypt the phrase, the encryption takes the phrase and encrypts it with the random encryption that was generated, the decryption does the opposite and gives back the original phrase.