Update EncryptionAndDecryptionJava
Ever wanted to generate a password that will be hard to crack? 

well here is a program that will generate a random encryption for the alphabet(all lowercase letters, using the ASCII table, where each letter from our alphabet will have a value that is between 2(min)-5(max) characters which can be changed, the code may be altered to be used however you want).
given a phrase/random string that is all lowercase and with no whitespaces. for example: "mynameisjeff".

The program does two things with the given phrase: encrypts the phrase or/and decrypts the phrase, the encryption function then takes the phrase and encrypts it with the random encryption that was generated.
The decryption function is given the encrypted version of the phrase and returns the original phrase(which is lowercase and without whitespaces).

min(parameter) has to be at least 2.