Create Algerbra_matrices
Given two matrices that are put into a file in the format at the bottom of the file, you can play around with your matrices in the main function. You can print the matrix, multiply two matrix, add two matrices to each other, get the transpose of a matrix or multiply/add a matrix with a scalar number.